requirements.txt
Chameleon==2.16
Jinja2==2.7.3
Mako==1.0.0
MarkupSafe==0.23
MySQL-python==1.2.5
PasteDeploy==1.5.2
Pygments==1.6
SQLAlchemy==0.9.7
WebOb==1.4
argparse==1.2.1
distribute==0.6.24
gevent==1.0.1
gevent-socketio==0.3.6
gevent-websocket==0.9.3
greenlet==0.4.3
gunicorn==19.1.1
pyramid==1.5.1
pyramid-chameleon==0.3
pyramid-debugtoolbar==2.2
pyramid-jinja2==2.3.3
pyramid-mako==1.0.2
redis==2.10.3
repoze.lru==0.6
transaction==1.4.3
translationstring==1.1
venusian==1.0
waitress==0.8.9
wsgiref==0.1.2
zope.deprecation==4.1.1
zope.interface==4.1.1
zope.sqlalchemy==0.7.5
